Move server log to server module.
