+++ server/src/main/java/io/atomix/catalog/server/request/AppendRequest.java
@@ -16,7 +16,7 @@
 package io.atomix.catalog.server.request;
 
 import io.atomix.catalog.client.request.AbstractRequest;
-import io.atomix.catalog.server.storage.Entry;
+import io.atomix.catalog.server.storage.entry.Entry;
 import io.atomix.catalyst.buffer.BufferInput;
 import io.atomix.catalyst.buffer.BufferOutput;
 import io.atomix.catalyst.serializer.SerializeWith;
