+++ server/src/main/java/io/atomix/catalog/server/state/FollowerState.java
@@ -15,20 +15,20 @@
  */
 package io.atomix.catalog.server.state;
 
-import io.atomix.catalog.server.request.VoteRequest;
-import io.atomix.catalog.server.RaftServer;
 import io.atomix.catalog.client.error.RaftError;
 import io.atomix.catalog.client.request.KeepAliveRequest;
 import io.atomix.catalog.client.request.RegisterRequest;
 import io.atomix.catalog.client.response.KeepAliveResponse;
 import io.atomix.catalog.client.response.RegisterResponse;
 import io.atomix.catalog.client.response.Response;
+import io.atomix.catalog.server.RaftServer;
 import io.atomix.catalog.server.request.AppendRequest;
 import io.atomix.catalog.server.request.PollRequest;
+import io.atomix.catalog.server.request.VoteRequest;
 import io.atomix.catalog.server.response.AppendResponse;
 import io.atomix.catalog.server.response.PollResponse;
 import io.atomix.catalog.server.response.VoteResponse;
-import io.atomix.catalog.server.storage.RaftEntry;
+import io.atomix.catalog.server.storage.entry.Entry;
 import io.atomix.catalog.server.util.Quorum;
 import io.atomix.catalyst.util.concurrent.Scheduled;
 
@@ -174,7 +174,7 @@ final class FollowerState extends ActiveState {
     // First, load the last log entry to get its term. We load the entry
     // by its index since the index is required by the protocol.
     long lastIndex = context.getLog().lastIndex();
-    RaftEntry lastEntry = lastIndex > 0 ? context.getLog().get(lastIndex) : null;
+    Entry lastEntry = lastIndex > 0 ? context.getLog().get(lastIndex) : null;
 
     final long lastTerm;
     if (lastEntry != null) {
