+++ server/src/main/java/io/atomix/catalog/server/state/PassiveState.java
@@ -21,9 +21,8 @@ import io.atomix.catalog.client.response.*;
 import io.atomix.catalog.server.RaftServer;
 import io.atomix.catalog.server.request.*;
 import io.atomix.catalog.server.response.*;
-import io.atomix.catalog.server.storage.ConfigurationEntry;
-import io.atomix.catalog.server.storage.Entry;
-import io.atomix.catalog.server.storage.RaftEntry;
+import io.atomix.catalog.server.storage.entry.ConfigurationEntry;
+import io.atomix.catalog.server.storage.entry.Entry;
 
 import java.util.ArrayDeque;
 import java.util.Queue;
@@ -110,7 +109,7 @@ class PassiveState extends AbstractState {
     }
 
     // If the previous entry term doesn't match the local previous term then reject the request.
-    try (RaftEntry entry = context.getLog().get(request.logIndex())) {
+    try (Entry entry = context.getLog().get(request.logIndex())) {
       if (entry == null || entry.getTerm() != request.logTerm()) {
         LOGGER.warn("{} - Rejected {}: Request log term does not match local log term {} for the same entry", context.getAddress(), request, entry != null ? entry.getTerm() : "unknown");
         return AppendResponse.builder()
@@ -138,14 +137,14 @@ class PassiveState extends AbstractState {
     if (!request.entries().isEmpty()) {
 
       // Iterate through request entries and append them to the log.
-      for (RaftEntry entry : (Iterable<RaftEntry>) request.entries()) {
+      for (Entry entry : request.entries()) {
         // If the entry index is greater than the last log index, skip missing entries.
         if (context.getLog().lastIndex() < entry.getIndex()) {
           context.getLog().skip(entry.getIndex() - context.getLog().lastIndex() - 1).append(entry);
           LOGGER.debug("{} - Appended {} to log at index {}", context.getAddress(), entry, entry.getIndex());
         } else {
           // Compare the term of the received entry with the matching entry in the log.
-          try (RaftEntry match = context.getLog().get(entry.getIndex())) {
+          try (Entry match = context.getLog().get(entry.getIndex())) {
             if (match != null) {
               if (entry.getTerm() != match.getTerm()) {
                 // We found an invalid entry in the log. Remove the invalid entry and append the new entry.
