+++ server/src/main/java/io/atomix/catalog/server/state/ServerStateMachine.java
@@ -15,10 +15,10 @@
  */
 package io.atomix.catalog.server.state;
 
-import io.atomix.catalog.server.storage.*;
 import io.atomix.catalog.server.StateMachine;
 import io.atomix.catalog.client.error.InternalException;
 import io.atomix.catalog.client.error.UnknownSessionException;
+import io.atomix.catalog.server.storage.entry.*;
 import io.atomix.catalyst.util.concurrent.ComposableFuture;
 import io.atomix.catalyst.util.concurrent.Context;
 import io.atomix.catalyst.util.concurrent.Futures;
