+++ server/src/main/java/io/atomix/catalog/server/storage/DescriptorException.java
