+++ server/src/main/java/io/atomix/catalog/server/storage/StorageException.java
