+++ server/src/main/java/io/atomix/catalog/server/storage/entry/QueryEntry.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.catalog.server.storage;
+package io.atomix.catalog.server.storage.entry;
 
 import io.atomix.catalog.client.Operation;
 import io.atomix.catalog.client.Query;
