+++ server/src/main/java/io/atomix/catalog/server/storage/entry/TimestampedEntry.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.catalog.server.storage;
+package io.atomix.catalog.server.storage.entry;
 
 import io.atomix.catalyst.buffer.BufferInput;
 import io.atomix.catalyst.buffer.BufferOutput;
@@ -21,56 +21,56 @@ import io.atomix.catalyst.serializer.Serializer;
 import io.atomix.catalyst.util.ReferenceManager;
 
 /**
- * Session entry.
+ * Timestamped entry.
  *
  * @author <a href="http://github.com/kuujo">Jordan Halterman</a>
  */
-public abstract class SessionEntry<T extends SessionEntry<T>> extends TimestampedEntry<T> {
-  private long session;
+public abstract class TimestampedEntry<T extends TimestampedEntry<T>> extends Entry<T> {
+  private long timestamp;
 
-  protected SessionEntry() {
+  protected TimestampedEntry() {
   }
 
-  protected SessionEntry(ReferenceManager<Entry<?>> referenceManager) {
+  protected TimestampedEntry(ReferenceManager<Entry<?>> referenceManager) {
     super(referenceManager);
   }
 
-  @Override
-  public long getAddress() {
-    return session;
+  /**
+   * Returns the entry timestamp.
+   *
+   * @return The entry timestamp.
+   */
+  public long getTimestamp() {
+    return timestamp;
   }
 
   /**
-   * Sets the session ID.
+   * Sets the entry timestamp.
    *
-   * @param session The session ID.
-   * @return The session entry.
+   * @param timestamp The entry timestamp.
+   * @return The entry.
    */
   @SuppressWarnings("unchecked")
-  public T setSession(long session) {
-    this.session = session;
+  public T setTimestamp(long timestamp) {
+    this.timestamp = timestamp;
     return (T) this;
-  }
-
-  /**
-   * Returns the session ID.
-   *
-   * @return The session ID.
-   */
-  public long getSession() {
-    return session;
   }
 
   @Override
   public void writeObject(BufferOutput buffer, Serializer serializer) {
     super.writeObject(buffer, serializer);
-    buffer.writeLong(session);
+    buffer.writeLong(timestamp);
   }
 
   @Override
   public void readObject(BufferInput buffer, Serializer serializer) {
     super.readObject(buffer, serializer);
-    session = buffer.readLong();
+    timestamp = buffer.readLong();
+  }
+
+  @Override
+  public String toString() {
+    return String.format("%s[index=%d, term=%d, timestamp=%d]", getClass().getSimpleName(), getIndex(), getTerm(), timestamp);
   }
 
 }
