+++ server/src/main/java/io/atomix/catalog/server/storage/entry/TypedEntryPool.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.catalog.server.storage;
+package io.atomix.catalog.server.storage.entry;
 
+import io.atomix.catalog.server.storage.StorageException;
 import io.atomix.catalyst.serializer.SerializationException;
 import io.atomix.catalyst.util.ReferenceManager;
 import io.atomix.catalyst.util.ReferencePool;
