+++ server/src/test/java/io/atomix/catalog/server/state/StateMachineTestCase.java
@@ -19,12 +19,12 @@ import io.atomix.catalog.server.StateMachine;
 import net.jodah.concurrentunit.ConcurrentTestCase;
 import io.atomix.catalog.client.Command;
 import io.atomix.catalog.client.Query;
-import io.atomix.catalog.server.storage.CommandEntry;
-import io.atomix.catalog.server.storage.KeepAliveEntry;
-import io.atomix.catalog.server.storage.QueryEntry;
-import io.atomix.catalog.server.storage.RegisterEntry;
+import io.atomix.catalog.server.storage.entry.CommandEntry;
+import io.atomix.catalog.server.storage.entry.KeepAliveEntry;
+import io.atomix.catalog.server.storage.entry.QueryEntry;
+import io.atomix.catalog.server.storage.entry.RegisterEntry;
 import io.atomix.catalog.client.session.Session;
-import io.atomix.catalog.server.storage.Entry;
+import io.atomix.catalog.server.storage.entry.Entry;
 import io.atomix.catalyst.serializer.Serializer;
 import io.atomix.catalyst.util.concurrent.Context;
 import io.atomix.catalyst.util.concurrent.SingleThreadContext;
