+++ server/src/test/java/io/atomix/catalog/server/storage/CleanerTest.java
