+++ server/src/test/java/io/atomix/catalog/server/storage/LogTest.java
@@ -15,6 +15,7 @@
  */
 package io.atomix.catalog.server.storage;
 
+import io.atomix.catalog.server.storage.entry.Entry;
 import io.atomix.catalyst.serializer.Serializer;
 import io.atomix.catalyst.serializer.ServiceLoaderTypeResolver;
 import org.testng.annotations.Test;
