+++ server/src/test/resources/META-INF/services/io.atomix.catalyst.serializer.CatalystSerializable
